Crate slack

Notes:

Structs except for RtmClient are derived from the slack api docs at: https://api.slack.com/

Optional fields in Slack structs represent fields that may not be present in the serialized form.

Currently fields in Slack structs are public, but in the future they may be wrapped in getters.

Usage: Implement an EventHandler to handle slack events and messages in conjunction with RtmClient.

Changelog:

Version 0.10.1: Massive overhaul, implement support for almost all of the bots api, stronger error handling and lots of tests. Thanks a ton to https://github.com/mthjones, see https://github.com/BenTheElder/slack-rs/pull/17 for the main overhaul.

Compatibility Changes: Methods that previously returned Result<String,Error> now return a typed Result<Some_Slack_Response_Type, Error>.

RtmClient::post_message now returns Result<api::chat::PostMessageResponse, Error>

RtmClient::delete_message now returns Result<api::chat::DeleteResponse, Error>

RtmClient::mark now returns Result<api::channels::MarkResponse, Error>

RtmClient::set_topic now returns Result<api::channels::SetTopicResponse, Error>

RtmClient::set_purpose now returns Result<api::channels::SetPurposeResponse, Error>

RtmClient::add_reaction_timestamp now returns Result<api::reactions::AddResponse, Error>

RtmClient::add_reaction_file now returns Result<api::reactions::AddResponse, Error>

RtmClient::add_reaction_file_comment now returns Result<api::reactions::AddResponse, Error>

RtmClient::update_message now returns Result<api::chat::UpdateResponse, Error>

RtmClient::im_open now returns Result<api::im::OpenResponse, Error>

RtmClient::channels_history now returns Result<api::channels::HistoryResponse, Error>

RtmClient::im_close now returns Result<api::im::CloseResponse, Error>

RtmClient::im_history now returns Result<api::im::HistoryResponse, Error>

RtmClient::im_list now returns Result<api::im::ListResponse, Error>

RtmClient::im_mark now returns Result<api::im::MarkResponse, Error>

Forthcoming releases will see the implementation of the remaining files.upload and some convenient helpers such as a message builder can be expected in a later release, and the Error::Api will expose Slack api error types more strongly in a forthcoming release.

Version 0.9.2: Add channels_history via https://github.com/jeehoonkang https://github.com/BenTheElder/slack-rs/pull/16

Version 0.9.1 -- With help from: https://github.com/mthjones, overhaul error handling and refactor, improve api support.

Introduced slack::error::Error

Added a number of bots api methods

Fixed bug where setPurpose called setTopic instead [!]

Version 0.8.3 -- Moved example to examples dir thanks to https://github.com/mthjones: https://github.com/BenTheElder/slack-rs/pull/9

Version 0.8.2 -- Fix https://github.com/BenTheElder/slack-rs/issues/8

Version 0.8.1 -- Add some web api methods, add methods to map names to ids.

TODO: expect the error type overhaul to be pushed back to 0.9.X

NOTE: Compatibility changes from 0.7.X include: RtmClient::new now takes the bot token/api_key and login, login_and_run do not.

Version 0.7.2 -- Bugfix via https://github.com/Farthen: https://github.com/BenTheElder/slack-rs/pull/6

Version 0.7.1 -- Cleaned up the api and json handling.

TODO: expect better error handling in 0.8.X

NOTE: Compatibility changes from 0.6.1 include: 'MessageHandler' is now 'EventHandler' and all of the slack data structs have been updated to match the api as closely as possible.

Version 0.6.1 -- Updated to stable rust.

Modules

api
Low-level, direct interface for the Slack Web API.

error

Structs

Attachment
The Slack Attachment object as found in richly-formatted messages.

Channel
The Slack Channel type.

Group
The Slack Group type.

Im
The Slack Im type.

RtmClient
The actual messaging client.

Team
The Team object as found in the rtm.start response.

User
The Slack User type.

Enums

Message
Represents a WebSocket message.

Traits

EventHandler
Implement this trait in your code to handle message events

Type Definitions

WsClient