Module slack::api::channels

Get info on your team's Slack channels, create or archive channels, invite users, set the topic and purpose, and mark a channel as read.

For more information, see Slack's API documentation.

Structs

AbridgedChannel

ArchiveResponse

CreateResponse

HistoryResponse

InfoResponse

InviteResponse

JoinResponse

KickResponse

LeaveResponse

ListResponse

MarkResponse

RenameResponse

SetPurposeResponse

SetTopicResponse

UnarchiveResponse

Functions

archive
Archives a channel.

create
Creates a channel.

history
Fetches history of messages and events from a channel.

info
Gets information about a channel.

invite
Invites a user to a channel.

join
Joins a channel, creating it if needed.

kick
Removes a user from a channel.

leave
Leaves a channel.

list
Lists all channels in a Slack team.

mark
Sets the read cursor in a channel.

rename
Renames a channel.

set_purpose
Sets the purpose for a channel.

set_topic
Sets the topic for a channel.

unarchive
Unarchives a channel.