Enum slack::error::Error

pub enum Error {
    Http(Error),
    WebSocket(Error),
    Utf8(FromUtf8Error),
    Url(UrlError),
    Json(Error),
    Api(String),
    Internal(String),
}

slack::Error represents errors that can happen while using the RtmClient

Variants
Http(Error)

Http client error

WebSocket(Error)

WebSocket connection error

Utf8(FromUtf8Error)

Error decoding websocket text frame Utf8

Url(UrlError)

Error parsing url

Json(Error)

Error decoding Json

Api(String)

Slack Api Error

Internal(String)

Errors that do not fit under the other types, Internal is for EG channel errors.