Trait slack::MessageHandler

pub trait MessageHandler {
    fn on_receive(&mut self, cli: &mut RtmClient, json_str: &str);
    fn on_ping(&mut self, cli: &mut RtmClient);
    fn on_close(&mut self, cli: &mut RtmClient);
    fn on_connect(&mut self, cli: &mut RtmClient);
}

Implement this trait in your code to handle message events

Required Methods

fn on_receive(&mut self, cli: &mut RtmClient, json_str: &str)

When a message is received this will be called with self, the slack client, and the json encoded string payload.

fn on_ping(&mut self, cli: &mut RtmClient)

Called when a ping is received; you do NOT need to handle the reply pong, but you may use this event to track the connection as a keep-alive.

fn on_close(&mut self, cli: &mut RtmClient)

Called when the connection is closed for any reason.

fn on_connect(&mut self, cli: &mut RtmClient)

Called when the connection is opened.