Crate slack

Notes:

Structs except for RtmClient are derived from the slack api docs at: https://api.slack.com/

Optional fields in Slack structs represent fields that may not be present in the serialized form.

Currently fields in Slack structs are public, but in 0.8.X onward they will likely be wrapped in getters.

Usage: Implement an EventHandler to handle slack events and messages in conjunction with RtmClient.

Changelog:

Version 0.8.2 -- Fix https://github.com/BenTheElder/slack-rs/issues/8

Version 0.8.1 -- Add some web api methods, add methods to map names to ids.

TODO: expect the error type overhaul to be pushed back to 0.9.X

NOTE: Compatibility changes from 0.7.X include: RtmClient::new now takes the bot token/api_key and login, login_and_run do not.

Version 0.7.2 -- Bugfix via https://github.com/Farthen: https://github.com/BenTheElder/slack-rs/pull/6

Version 0.7.1 -- Cleaned up the api and json handling.

TODO: expect better error handling in 0.8.X

NOTE: Compatibility changes from 0.6.1 include: 'MessageHandler' is now 'EventHandler' and all of the slack data structs have been updated to match the api as closely as possible.

Version 0.6.1 -- Updated to stable rust.

Structs

Channel
See the slack api docs at: https://api.slack.com/

Group
See the slack api docs at: https://api.slack.com/

Im
See the slack api docs at: https://api.slack.com/

Purpose
See the slack api docs at: https://api.slack.com/

RtmClient
The actual messaging client.

RtmStart
See the slack api docs at: https://api.slack.com/

SelfData
See the slack api docs at: https://api.slack.com/

Team
See the slack api docs at: https://api.slack.com/

Topic
See the slack api docs at: https://api.slack.com/

User
See the slack api docs at: https://api.slack.com/

Enums

Message
Represents a WebSocket message.

Traits

EventHandler
Implement this trait in your code to handle message events

Type Definitions

WsClient