Struct slack::RtmClient

pub struct RtmClient {
    // some fields omitted
}

The actual messaging client.

Methods

impl RtmClient

fn new() -> RtmClient

Creates a new empty client.

fn get_outs(&self) -> Option<Sender<Message>>

fn get_name(&self) -> String

Returns the name of the bot/user connected to the client. Only valid after login.

fn get_id(&self) -> String

Returns the id of the bot/user connected to the client. Only valid after login.

fn get_team<'a>(&'a self) -> &'a Team

Returns the Team struct of the bot/user connected to the client. Only valid after login.

fn get_msg_uid(&self) -> isize

Returns a unique identifier to be used in the 'id' field of a message sent to slack.

fn send(&mut self, s: &str) -> Result<(), String>

Allows sending a json string message over the websocket connection. Note that this only passes the message over a channel to the Messaging task, and therfore a succesful return value does not mean the message has been actually put on the wire yet. Note that you will need to form a valid json reply yourself if you use this method, and you will also need to retrieve a unique id for the message via RtmClient.get_msg_uid() Only valid after login.

fn send_message(&self, chan: &str, msg: &str) -> Result<(), String>

Allows sending a textual string message over the websocket connection, to the requested channel id. Ideal usage would be EG: extract the channel in on_receive and then send back a message to the channel. Note that this only passes the message over a channel to the Messaging task, and therfore a succesful return value does not mean the message has been actually put on the wire yet. This method also handles getting a unique id and formatting the actual json sent. Only valid after login.

fn login(&mut self, token: &str) -> Result<(WsClient, Receiver<Message>), String>

Logs in to slack. Call this before calling run. Alternatively use login_and_run

fn run<T: MessageHandler>(&mut self, handler: &mut T, client: WsClient, rx: Receiver<Message>) -> Result<(), String>

Runs the message receive loop

fn login_and_run<T: MessageHandler>(&mut self, handler: &mut T, token: &str) -> Result<(), String>

Runs the main loop for the client after logging in to slack, returns an error if the process fails at an point, or an Ok(()) on succesful close. Takes a MessageHandler (implemented by the user) to call events handlers on. once the first on_receive() or on_ping is called on the MessageHandler, you can soon the 'Only valid after login' methods are safe to use. Sending is run in a thread in parallel while the receive loop runs on the main thread. Both loops should end on return. Sending should be thread safe as the messages are passed in via a channel in RtmClient.send and RtmClient.send_message